+++ injection_scripts/selector_query.js
@@ -66,16 +66,27 @@ if (!window.WendigoQuery) {
 
         _parseSelectorType(selector) {
             if (typeof(selector) === "string") {
-                return this._parseStringSelector(selector);
+                if (selector.length === 0) return null;
+                if (this._isXPathQuery(selector)) return this.selectorTypes.xpath;
+                else return this.selectorTypes.css;
             } else if (typeof(selector) === "object") {
                 return this.selectorTypes.domElement;
             } else return null;
         },
 
-        _parseStringSelector(selector) {
-            if (selector.length === 0) return null;
-            if (selector[0] === "/") return this.selectorTypes.xpath;
-            else return this.selectorTypes.css;
+        // NOTE: Duplicate of utils.isXPathQuery
+        _isXPathQuery(s) {
+            if (s[0] === '/') return true;
+            if (/^.\./.test(s)) return true;
+            const axisSplit = s.split("::");
+            if (axisSplit.length > 1) {
+                const validAxis = ["ancestor", "ancestor-or-self", "attribute", "child", "descendant", "descendant-or-self",
+                    "following", "following-sibling", "namespace", "parent", "preceding", "preceding-sibling", "self"
+                ];
+                const axe = axisSplit[0];
+                if (validAxis.indexOf(axe) !== -1) return true;
+            }
+            return false;
         }
     };
 }